add WasmBool, replace many int -> size_t

+++ src/wasm-array.h
@@ -50,7 +50,7 @@
 
 #define WASM_DESTROY_ARRAY_AND_ELEMENTS(allocator, v, name) \
   {                                                         \
-    int i;                                                  \
+    size_t i;                                               \
     for (i = 0; i < (v).size; ++i)                          \
       wasm_destroy_##name(allocator, &((v).data[i]));       \
     wasm_destroy_##name##_array(allocator, &(v));           \
